requirements.txt
aiocache==0.12.3
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
alembic==1.14.0
annotated-types==0.7.0
anthropic==0.45.2
anyio==4.8.0
appdirs==1.4.4
APScheduler==3.10.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.1.0
Authlib==1.4.1
av==12.3.0
backoff==2.2.1
bcrypt==4.2.0
beautifulsoup4==4.13.3
bidict==0.23.1
bitarray==3.0.0
black==24.8.0
blinker==1.9.0
boto3==1.35.53
botocore==1.35.99
build==1.2.2.post1
cachetools==5.5.1
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
chroma-hnswlib==0.7.6
chromadb==0.6.2
click==8.1.8
colbert-ai==0.2.21
colorama==0.4.6
colorclass==2.2.2
coloredlogs==15.0.1
compressed_rtf==1.0.6
cryptography==44.0.0
ctranslate2==4.5.0
dataclasses-json==0.6.7
datasets==3.2.0
decompyle3==3.9.2
defusedxml==0.7.1
Deprecated==1.2.18
dill==0.3.8
distro==1.9.0
dnspython==2.7.0
docker==7.1.0
docx2txt==0.8
duckduckgo_search==7.3.1
durationpy==0.9
easygui==0.98.3
ebcdic==1.1.1
ecdsa==0.19.0
einops==0.8.0
emoji==2.14.1
et_xmlfile==2.0.0
eval_type_backport==0.2.2
Events==0.5
extract-msg==0.53.1
fake-useragent==1.5.1
fastapi==0.115.7
faster-whisper==1.0.3
filelock==3.17.0
filetype==1.2.0
Flask==3.1.0
Flask-Cors==5.0.0
flatbuffers==25.1.24
fonttools==4.56.0
fpdf2==2.8.2
frozenlist==1.5.0
fs==2.4.16
fsspec==2024.9.0
ftfy==6.2.3
gcp-storage-emulator==2024.8.3
git-python==1.0.3
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.6
google-api-core==2.24.1
google-api-python-client==2.160.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud-core==2.4.1
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-generativeai==0.7.2
google-resumable-media==2.7.2
googleapis-common-protos==1.63.2
greenlet==3.1.1
grpcio==1.67.1
grpcio-status==1.62.3
grpcio-tools==1.62.3
h11==0.14.0
h2==4.2.0
hpack==4.1.0
html5lib==1.1
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.28.1
humanfriendly==10.0
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.4.0
importlib_resources==6.5.2
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
kubernetes==32.0.0
langchain==0.3.7
langchain-community==0.3.7
langchain-core==0.3.34
langchain-text-splitters==0.3.6
langdetect==1.0.9
langfuse==2.44.0
langsmith==0.1.147
lark==1.1.9
ldap3==2.9.1
lxml==5.3.0
Mako==1.3.9
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
mmh3==5.1.0
monotonic==1.6
moto==5.0.28
mpmath==1.3.0
msoffcrypto-tool==5.4.2
multidict==6.1.0
multiprocess==0.70.16
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.11.1.3
nltk==3.9.1
numpy==1.26.4
oauthlib==3.2.2
olefile==0.47
oletools==0.60.2
onnxruntime==1.20.1
open-webui==0.5.10
openai==1.61.1
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
opensearch-py==2.7.1
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-instrumentation==0.48b0
opentelemetry-instrumentation-asgi==0.48b0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
opentelemetry-util-http==0.48b0
orjson==3.10.15
overrides==7.7.0
packaging==23.2
pandas==2.2.3
passlib==1.7.4
pathspec==0.12.1
pcodedmp==1.2.6
peewee==3.17.8
peewee-migrate==1.12.2
pgvector==0.3.5
pillow==11.1.0
platformdirs==4.3.6
pluggy==1.5.0
portalocker==2.10.1
posthog==3.11.0
primp==0.12.0
propcache==0.2.1
proto-plus==1.26.0
protobuf==4.25.6
psutil==6.1.1
psycopg2-binary==2.9.9
py-partiql-parser==0.6.1
pyarrow==19.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyclipper==1.3.0.post6
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.9.2
pydantic-settings==2.7.1
pydantic_core==2.23.4
pydub==0.25.1
Pygments==2.19.1
PyJWT==2.10.1
pymdown-extensions==10.14.2
pymilvus==2.5.0
pymongo==4.11
PyMySQL==1.1.1
PyNaCl==1.5.0
pypandoc==1.13
pyparsing==3.2.1
pypdf==4.3.1
PyPika==0.48.9
pypiwin32==223
pyproject_hooks==1.2.0
pyreadline3==3.5.4
pytest==8.3.4
pytest-docker==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.11.2
python-iso639==2025.1.28
python-jose==3.3.0
python-magic==0.4.27
python-multipart==0.0.18
python-oxmsg==0.0.2
python-pptx==1.0.0
python-socketio==5.11.3
pytube==15.0.0
pytz==2025.1
pywin32==308
pyxlsb==1.0.10
PyYAML==6.0.2
qdrant-client==1.12.2
rank-bm25==0.2.2
RapidFuzz==3.12.1
rapidocr-onnxruntime==1.3.24
rblx-open-cloud==2.2.4
red-black-tree-mod==1.22
redis==5.2.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
responses==0.25.6
rich==13.9.4
rsa==4.9
RTFDE==0.1.2
s3transfer==0.10.4
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.1
sentence-transformers==3.3.1
sentencepiece==0.2.0
shapely==2.0.7
shellingham==1.5.4
simple-websocket==1.1.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soundfile==0.13.1
soupsieve==2.6
spark-parser==1.8.9
SQLAlchemy==2.0.32
starlette==0.45.3
sympy==1.13.1
tenacity==9.0.0
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.21.0
torch==2.6.0
tqdm==4.67.1
transformers==4.48.3
typer==0.15.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.2
ujson==5.10.0
uncompyle6==3.9.2
unstructured==0.16.11
unstructured-client==0.28.1
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.30.6
validators==0.34.0
watchfiles==1.0.4
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
websockets==14.2
Werkzeug==3.1.3
win_unicode_console==0.5
wrapt==1.17.2
wsproto==1.2.0
xdis==6.1.2
xlrd==2.0.1
XlsxWriter==3.2.2
xmltodict==0.14.2
xxhash==3.5.0
yarl==1.18.3
youtube-transcript-api==0.6.3
zipp==3.21.0
